% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\docType{class}
\name{acs-class}
\alias{acs-class}
\alias{acs.colnames,acs-method}
\alias{acs.units,acs-method}
\alias{-,acs,acs-method}
\alias{-,acs,numeric-method}
\alias{-,numeric,acs-method}
\alias{/,acs,acs-method}
\alias{/,acs,numeric-method}
\alias{/,numeric,acs-method}
\alias{[<-,acs-method}
\alias{[,acs-method}
\alias{*,acs,acs-method}
\alias{*,acs,numeric-method}
\alias{*,numeric,acs-method}
\alias{+,acs,acs-method}
\alias{+,acs,numeric-method}
\alias{+,numeric,acs-method}
\alias{estimate,acs-method}
\alias{modified,acs-method}
\alias{show,acs-method}
\alias{span,acs-method}
\alias{standard.error,acs-method}
\alias{summary,acs-method}
\alias{estimate}
\alias{modified}
\alias{span}
\alias{standard.error}
\alias{acs.colnames}
\alias{acs.units}
\alias{dim.acs}
\alias{length.acs}
\alias{acs.colnames<-}
\alias{acs.units<-}
\alias{is.acs}
\alias{span<-}
\alias{apply}
\alias{acs.colnames<-,acs-method}
\alias{acs.units<-,acs-method}
\alias{span<-,acs-method}
\alias{apply,acs-method}
\title{Class \code{"acs"}}
\description{
The acs class provides a convenient wrapper for demographic data from the
U.S. Census, especially the American Community Survey.  Estimates and
standard errors are kept together, along with geographic information and
metadata necessary to manipulate and analyze data in this form.
}
\section{Objects from the Class}{


acs objects can be created by calls of the form \code{new("acs", ...)}, or
through helper functions provided by the package (currently \code{read.acs}
and \code{acs.fetch}), or from the output of subsetting or other calls on
existing acs objects.  Once created, acs objects can be manipulated through
new methods to deal appropriately with common analytical tasks such as
combining subgroups or geographies, mathematical operations on estimates,
and computing (and plotting) confidence intervals.
}

\examples{

showClass("acs")
# load some data from the ACS
data(kansas09)
str(kansas09)

# access slots
endyear(kansas09)
span(kansas09)
estimate(kansas09)[1:5,1:5]
standard.error(kansas09[1:5,1:5])


# subset
kansas09[1:4,6:9]

# more complicated subsets
kansas09[c("Linn County, Kansas", "Wilson County, Kansas") ,
   grep(pattern="21.years", x=acs.colnames(kansas09))]

# addition on estimates and errors
kansas09[1:4,25]+kansas09[1:4,49]

# can even multiply and divide
# males per female, by county
kansas09[1:4,2]/kansas09[1:4,26]

# (males<5 plus females<5) * 12
(kansas09[7,3]+kansas09[7,27]) * 12

# some replacement: males<5 as a percentage of all males
kansas09[,3]=kansas09[,3]/kansas09[,2]

}
\author{
Ezra Haber Glenn \email{eglenn@mit.edu}
}
\keyword{classes}
