% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-package.R
\docType{package}
\name{acs-package}
\alias{acs}
\alias{acs-package}
\title{acs: Tools for Working with IPUMS ACS Data}
\description{
Tools for working with IPUMS ACS data.
}
\author{
\strong{Maintainer}: Ray Wright \email{raywright88@gmail.com}

}
\keyword{internal}
