% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{class}
\name{acs.lookup-class}
\alias{acs.lookup-class}
\alias{is.acs.lookup}
\alias{results}
\alias{+,acs.lookup,acs.lookup-method}
\alias{c,acs.lookup-method}
\alias{endyear,acs.lookup-method}
\alias{show,acs.lookup-method}
\alias{[,acs.lookup-method}
\alias{results,acs.lookup-method}
\alias{span,acs.lookup-method}
\title{Class \code{"acs.lookup"}}
\description{
A new class to hold the results of calls to \code{acs.lookup}, typically
for possible modification and then passing to calls to \code{acs.fetch}
(using the "variable=" argument).
}
\section{Objects from the Class}{
 Objects can be created by calls of the
form \code{new("acs.lookup", ...)}, but more typically will be created as
output from calls to \code{acs.lookup}.
}

\examples{

showClass("acs.lookup")

}
\seealso{
\code{\link{acs.lookup}}
}
\author{
Ezra Haber Glenn \email{eglenn@mit.edu}
}
\keyword{classes}
