% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{acs_clean}
\alias{acs_clean}
\title{Clean IPUMS ACS data}
\usage{
acs_clean(data)
}
\arguments{
\item{data}{Data frame of IPUMS ACS data.}
}
\value{
Data frame.
}
\description{
Currently, this only cleans/recodes a selection of variables.
}
