% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api.key.migrate}
\alias{api.key.migrate}
\title{After updating the acs package, installs an archived API key from a
previous installation.}
\usage{
api.key.migrate()
}
\value{
Migrates the key (if found) and exits silently; return an error if
no archived key is found.
}
\description{
The \code{acs.fetch} function requires an "API key" to use when downloading
data from the US Census API.  Rather than pass this rather long string to
the function each time, users can save the key as part of the package
installation, using the \code{api.key.install} function. Once installed, an
api key is saved on the system and available for use in future sessions.
(To replace a key, simply call the function again with the new key.)
}
\details{
During the update process, this key may be lost or left in the wrong
location.  A call to \code{api.key.migrate()} can help restore an archived
key, if found.
}
\references{
To request an API key, see
\url{http://www.census.gov/developers/}
}
\seealso{
\code{\link{acs.fetch}} \code{\link{api.key.install}}
}
\author{
Ezra Haber Glenn \email{eglenn@mit.edu}
}
