% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{Consumer Price Index data (1913-2015).}
\format{A named vector containing 103 observations, one for each year from
1913 through 2015.}
\source{
\url{http://www.bls.gov/cpi/}
}
\description{
Contains data on the Consumer Price Index for All Urban Consumers (CPI-U)
for the years 1913-2015 from the U.S. Bureau of Labor Statistics.  Used by
the acs package for currency conversion functions.  Scaled for base (100)
to be 1982-84.
}
\seealso{
\code{\link{currency.year}}

\code{\link{currency.convert}}
}
\keyword{datasets}
