% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_degree}
\alias{cw_degree}
\title{University degree crosswalk}
\format{A data frame with two variables:
\describe{
\item{\code{degree_code}}{Code of degree.}
\item{\code{degree_name}}{Name of degree.}
}}
\usage{
cw_degree
}
\description{
University degree crosswalk
}
\keyword{datasets}
