% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_metro}
\alias{cw_metro}
\title{Metro area crosswalk}
\format{A data frame with two variables:
\describe{
\item{\code{msa_code}}{Code of metro area.}
\item{\code{msa_name}}{Name of metro area.}
}}
\usage{
cw_metro
}
\description{
Metro area crosswalk
}
\keyword{datasets}
