% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_occ}
\alias{cw_occ}
\title{Occupation crosswalk}
\format{A data frame with three variables:
\describe{
\item{\code{occ_code}}{Code of occupation.}
\item{\code{occ_name}}{Name of occupation.}
\item{\code{occ_cat_name}}{Name of occupation major category.}
}}
\usage{
cw_occ
}
\description{
Occupation crosswalk
}
\keyword{datasets}
