% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{db_list_acs}
\alias{db_list_acs}
\title{List variables in IPUMS ACS database table}
\usage{
db_list_acs(file_db)
}
\arguments{
\item{file_db}{Path of database file to read.}
}
\value{
Character vector of variable names from database table.
}
\description{
List variables in IPUMS ACS database table
}
