% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{db_read_acs}
\alias{db_read_acs}
\title{Read database table.}
\usage{
db_read_acs(file_db, years, vars)
}
\arguments{
\item{file_db}{Path of database file to read.}

\item{years}{Years in data to query (where).}

\item{vars}{Variables in data to query (select).}
}
\value{
Data frame.
}
\description{
Read database table.
}
