% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_db_table}
\alias{write_db_table}
\title{Write database table from CSV file}
\usage{
write_db_table(file_data, file_db)
}
\arguments{
\item{file_data}{Path of CSV file of IPUMS ACS data.}

\item{file_db}{Path of database to create.}
}
\description{
This will overwrite the existing database table with the same name.
}
