% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{db_write_acs}
\alias{db_write_acs}
\title{Write IPUMS ACS database table}
\usage{
db_write_acs(file_data, file_db)
}
\arguments{
\item{file_data}{Path of CSV file of IPUMS ACS data.}

\item{file_db}{Path of database to create.}
}
\description{
Write database table from IPUMS ACS CSV file. This will overwrite the
existing database file with the same name.
}
