\name{fips.state}
\alias{fips.state}
\alias{fips.county}
\alias{fips.county.subdivision}
\alias{fips.place}
\alias{fips.school}
\alias{fips.american.indian.area}
\docType{data}
\title{
FIPS codes and geographic names for use in searching and creating
geographies in the acs package.
}
\description{
FIPS codes and geographic names for use in searching and creating
geographies in the acs package.  (Used internally.)
}
\usage{data(fips.state)}
\format{Each table is a dataframe containing FIPS codes and names from
  the US Census.} 

\source{

  State: http://www.census.gov/geo/reference/ansi_statetables.html
  
County: http://www.census.gov/geo/www/codes/county/download.html

County Subdivision:
http://www.census.gov/geo/www/codes/cousub/download.html

Place: http://www.census.gov/geo/www/codes/place/download.html

School: http://www.census.gov/geo/www/codes/sd/

American Indian Area:  http://www.census.gov/geo/www/codes/aia/
}


\keyword{datasets}
